bugfixes and improvements

+++ src/vm16.h
@@ -134,7 +134,7 @@ uint32_t vm16_write_mem(vm16_t *C, uint16_t addr, uint16_t num, uint16_t *p_buff
 /*
 ** Read memory block and convert the data to ASCII characters (screen memory)
 */
-uint32_t vm16_get_ascii(vm16_t *C, uint16_t addr, uint16_t num, char *p_buffer);
+uint16_t vm16_get_ascii(vm16_t *C, uint16_t addr, uint16_t num, char *p_buffer);
 
 /*
 ** Read value from VM memory
